Extracted componentable parts from srt-test-relay
